Correction that fixes problem of no fire after year 1
The two new flags identifying when passive and active crown fires
occur needed to be integers for the if statements to work correctly.

To facilitate my debugging, I reverted some aesthetic code mods to look
more like @jkshuman's read_lightning_pr561 branch
(https://github.com/jkshuman/fates/tree/read_lightning_pr561) so as to
minimize unnecessary diffs when typing "git diff read_lightning_pr561".
These mods will likely make it back into this branch (active_crown_fire)
with later updates.

+++ main/EDTypesMod.F90
@@ -344,8 +344,8 @@ module EDTypesMod
 
      ! FIRE
      real(r8) ::  fraction_crown_burned                  ! proportion of crown affected by fire:-
-     real(r8) ::  active_crown_fire_flg                  ! flag for active crown fire ignition
-     real(r8) ::  passive_crown_fire_flg                 ! flag for passive crown fire ignition (1=ignition)
+     integer ::  active_crown_fire_flg                   ! flag for active crown fire ignition
+     integer ::  passive_crown_fire_flg                  ! flag for passive crown fire ignition (1=ignition)
      real(r8) ::  cambial_mort                           ! probability that trees dies due to cambial char 
                                                          ! (conditional on the tree being subjected to the fire)
      real(r8) ::  crownfire_mort                         ! probability of tree post-fire mortality
